Chore: Bump up ESP-IDF v5.1 Signed-off-by: Jeff Yoon <jeyoon@outlook.com>

+++ product/actuator/main/actuator_mqtt_task.c
@@ -626,7 +626,7 @@ static char *gen_devinfo_resp(void) {
   esp_netif_get_ip_info(esp_netif_get_handle_from_ifkey("ETH_DEF"), &ip_info);
   esp_ip4addr_ntoa(&ip_info.ip, ip_addr, sizeof(ip_addr));
 
-  snprintf(free_mem, sizeof(free_mem), "%ld", xPortGetFreeHeapSize());
+  snprintf(free_mem, sizeof(free_mem), "%d", (int)xPortGetFreeHeapSize());
   syscfg_get(SYSCFG_I_FWVERSION, SYSCFG_N_FWVERSION, fw_version, sizeof(fw_version));
   syscfg_get(SYSCFG_I_RECONNECT, SYSCFG_N_RECONNECT, reconnect, sizeof(reconnect));
 
@@ -1359,6 +1359,5 @@ void create_mqtt_task(void) {
 
 #endif
 
-  xTaskCreatePinnedToCore(mqtt_task, MQTT_TASK_NAME, 4096, NULL, (tskIDLE_PRIORITY + 5),
-                          &mqtt_task_handle, 1);
+  xTaskCreatePinnedToCore(mqtt_task, MQTT_TASK_NAME, 4096, NULL, (tskIDLE_PRIORITY + 5), &mqtt_task_handle, 1);
 }
